+++ demisto_sdk/commands/split/tests/ymlsplitter_test.py
@@ -2,6 +2,7 @@ import base64
 import os
 from pathlib import Path
 
+from TestSuite.test_tools import ChangeCWD
 from demisto_sdk.commands.common.configuration import Configuration
 from demisto_sdk.commands.common.constants import DEFAULT_IMAGE_BASE64
 from demisto_sdk.commands.common.handlers import YAML_Handler
@@ -12,7 +13,6 @@ yaml = YAML_Handler()
 
 
 def test_extract_long_description(tmpdir):
-
     # Test when script
     extractor = YmlSplitter(input=f'{git_path()}/demisto_sdk/tests/test_files/script-test_script.yml',
                             output='', file_type='script', no_demisto_mock=False,
@@ -30,7 +30,6 @@ def test_extract_long_description(tmpdir):
 
 
 def test_extract_image(tmpdir):
-
     # Test when script
     extractor = YmlSplitter(input=f'{git_path()}/demisto_sdk/tests/test_files/script-test_script.yml',
                             output='', file_type='script')
@@ -130,6 +129,18 @@ def test_get_output_path():
                             output=out)
     res = extractor.get_output_path()
     assert res == Path(out + "/Zoom")
+
+
+def test_get_output_path_relative(repo):
+    pack = repo.create_pack()
+    integration = pack.create_integration()
+
+    with ChangeCWD(repo.path):
+        extractor = YmlSplitter(input=integration.yml.rel_path, file_type='integration')
+
+    output_path = extractor.get_output_path()
+    assert output_path.is_absolute()
+    assert output_path.relative_to(pack.path) == Path(integration.path).relative_to(pack.path)
 
 
 def test_get_output_path_empty_output():